Set the log level in operand.
Sidecar containers have the log level set through the
${LOG_LEVEL} placeholder, as usual. That placeholder
doesn't work for vSphere CSI driver and syncer, though.

First of all, the Deployment controller from library-go
replaces ${LOG_LEVEL} with a number representing the log
level. However, both the CSI driver and syncer containers
expect the log level to be set through the environment
variable X_CSI_LOG_LEVEL as a string, as in Debug, Trace,
etc.

Another problem is that the CSI driver uses logrus instead
of klog; and numbers representing log levels have different
meanings there. For instance, the verbosity value for Debug
in klog is 4, while in logrus it's 5.

In order to overcome these differences, this patch takes the
following approach: when CR.Spec.LogLevel is Debug or higher,
we set the X_CSI_DEBUG environment variable in the CSI driver
and syncer to a positive value.