Deploy multiple replicas of CSI controller
Currently, we deploy only one replica of the CSI Controller
Service Deployment. This apparently leads to the operator
going unavailable when draining the node where the Deployment
is running on. Also, since we use the RollingUpdate strategy,
Deployment rollous get stuck in single-node clusters because
the new pod waits for a node with free ports, which never
happens because the old pod is waiting for the new pod to start.

This patch uses the hook introduced in openshift/library-go#1056
to run multiple replicas of the operand, i.e., 1 in single-node
control-planes or 2 in multiple-node control-planes. In addition
to that, we set maxSurge=0 and maxUnavailable=1 in the RollingUpdate
strategy to make sure pods are drained one at a time and to prevent
more than one pod (same operand) running at the same time in the
node, since we use nodePorts and hostNetwork in this operator.
Also, we set anti-affinity rules to avoid two replicas running
on the same master node.