reset operator metrics on successful sync
If operator creates a metric indicating error the metric is never
removed and reported even if user resolved the issue.

For example setting incorrect password for vsphere connection degrades
the operator state and emits a metric when periodic sync is performed.

If user then corrects the password and vsphere client succeeds this
metric is persisted.

This patch should ensure we reset the metric if we don't detect any
known negative state (defined in blockUpgradeOrDegradeCluster())