Add BlockUpgradeOrDegrade check result
This result depends on whether the driver is installed - if so, then
degrade the clusters, if it's not installed (or the community version is
installed) then just mark the cluster un-upgradeable.

This was actualy the behavior of "BlockUpgrade" check result, which was
confusing. A subsequent patch will need to mark the cluster unupgradeable
without degrading the cluster.