add check for max volumes per node
We need to check versions of all ESXI hosts in the cluster and if we
detect that users set a custom volume attachment limit that is
incorrect we degrade the cluster.

Incorrect value is any value above 59 if any of the vSphere hosts in a
cluster is not on ESXI version 8 or higher

In case the maxAllowedBlockVolumesPerNode field is not set it will
return 0, which is not valid and we need to use default.